ci/github-script/commits: split review function into separate file
This allows re-using postReview in the next commit.
